requirements.txt
django==3.1
psycopg2-binary==2.8.5 
django-allauth==0.42.0
djangorestframework==3.11.1
django-cors-headers==3.4.0